requirements.txt
# Polymarket BTC 15-Min Bot Dependencies

# Polymarket CLOB client
py-clob-client>=0.34.5

# Web3 / Crypto
web3>=6.0.0
eth-account>=0.10.0

# HTTP / WebSocket
aiohttp>=3.9.0
websockets>=12.0
requests>=2.31.0

# Data processing
numpy>=1.26.0
pandas>=2.1.0

# Monitoring
prometheus-client>=0.20.0

# Telegram alerts
python-telegram-bot>=20.7

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging
structlog>=23.2.0

# Scheduling
apscheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
